backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP Client
httpx==0.26.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LLM (Ollama - local, free)
ollama==0.1.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
